backend/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.11",
    "router>=0.1",
    "routers>=0.10.1",
    "uvicorn>=0.34.0",
    "sqlalchemy>=2.0.30",
    "python-jose>=3.4.0",
    "passlib>=1.7.4",
    "PyJWT>=2.8.0",
    "aiomysql>=0.2.0",
    "pymysql>=1.1.0",
    "cryptography>=42.0.0",
    "bcrypt>=4.0.1",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.1",
    "pydantic_settings>=2.8.1",
    "pyyaml>=6.0.2",
    "mahjong>=1.2.1",
    "websockets>=15.0.1",
]
